+++ .storybook/main.ts
@@ -1,28 +1,27 @@
-import type { StorybookConfig } from '@storybook/react-webpack5';
+import type { StorybookConfig } from '@storybook/react-vite';
 import path from 'path';
 
 const config: StorybookConfig = {
   stories: [
     '../src/**/*.stories.@(ts|tsx|mdx)',
-    '../docs/**/*.stories.@(mdx)'
+    '../docs/**/*.stories.mdx'
   ],
   addons: [
+    '@storybook/addon-docs',
     '@storybook/addon-essentials',
     '@storybook/addon-a11y',
-    'storybook-addon-designs',
+    '@storybook/addon-viewport',
+    '@storybook/addon-designs',
     '@chromatic-com/storybook'
   ],
   framework: {
-    name: '@storybook/react-webpack5',
+    name: '@storybook/react-vite',
     options: {}
-  },
-  core: {
-    builder: '@storybook/builder-webpack5'
   },
   docs: {
     defaultName: 'Docs'
   },
-  webpackFinal: async (config: any) => {
+  viteFinal: async (config: any) => {
     config.resolve = config.resolve || {};
     config.resolve.alias = {
       ...(config.resolve.alias || {}),