+++ .storybook/preview.ts
@@ -1,8 +1,7 @@
 import React from 'react';
-import type { Preview } from '@storybook/react';
 import '../src/index.css';
 
-const preview: Preview = {
+const preview = {
   parameters: {
     controls: {
       matchers: {